+++ tests/src/Interop/CMakeLists.txt
@@ -55,7 +55,7 @@ add_subdirectory(StringMarshalling/AnsiBSTR)
 add_subdirectory(StringMarshalling/VBByRefStr)
 add_subdirectory(MarshalAPI/FunctionPointer)
 add_subdirectory(MarshalAPI/IUnknown)
-add_subdirectory(MarshalAPI/NativeLibrary)
+add_subdirectory(NativeLibrary)
 add_subdirectory(SizeConst)
 add_subdirectory(DllImportAttribute/ExeFile)
 add_subdirectory(DllImportAttribute/FileNameContainDot)
